Actually make the version bump run before packaging
I read the travis docs, and it seems that I can't reuse the storage
between different jobs, and that I'm using the build matrix incorrectly.
the build matrix is to run the same tests under the different
environments, which we will use for tf vs torch testing.  But not for
different branches.